// commerceos/libs/apps/products/src/lib/shared/enums/editor.enum.ts
export enum ActiveView {
  ProductList = 'list',
  ProductAdd = 'add',
  ProductEdit = 'edit',
  ProductsImport = 'import',
  CollectionAdd = 'add-collection',
  CollectionEdit = 'edit-collection',
}

export enum NavbarControlType {
  Text = 'text',
  Link = 'link',
  Menu = 'menu',
  Divider = 'divider',
  Custom = 'custom',
  CustomElement = 'customElement',
}

export enum NavbarControlPosition {
  Left = 'left',
  Center = 'center',
  Right = 'right',
}